更改插件版本,指令模块增加管理员权限

+++ src/main/java/world/thek/controller/OrderController.java
@@ -5,6 +5,7 @@ import net.mamoe.mirai.event.SimpleListenerHost;
 import net.mamoe.mirai.event.events.MessageEvent;
 import net.mamoe.mirai.message.data.MessageChain;
 import net.mamoe.mirai.message.data.MessageChainBuilder;
+import world.thek.config.ConfigData;
 import world.thek.entity.Order;
 import world.thek.util.ConstantUtil;
 import world.thek.util.FileUtil;
@@ -20,26 +21,32 @@ public class OrderController extends SimpleListenerHost {
 
     @EventHandler
     public void order(MessageEvent event) throws IOException {
+        long userId = event.getSender().getId();
         String code = event.getMessage().serializeToMiraiCode();
         String[] split = code.split("\\s+");
         String order = split[0];
 
         Map<String,String> orderMap = FileUtil.read(ConstantUtil.ORDER_FILENAME);
 
-        if (orderMap.get(order) != null) {
-            String value = orderMap.get(order);
-            event.getSubject().sendMessage(value);
-        }
-
         //添加指令
         if (Order.ORDER_ADD.equals(order)) {
-            String key = split[1];
-            String value = split[2];
-            if (orderMap.get(key) != null) {
-                event.getSubject().sendMessage("指令已存在！");
+            if (userId == ConfigData.INSTANCE.getOwner()) {
+                String key = split[1];
+                String value = split[2];
+                if (orderMap.get(key) != null) {
+                    event.getSubject().sendMessage("指令已存在！");
+                } else {
+                    FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
+                    event.getSubject().sendMessage("添加指令: "+key);
+                }
             } else {
-                FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
-                event.getSubject().sendMessage("添加指令: "+key);
+                MessageChainBuilder messages = new MessageChainBuilder();
+                messages.append("当前用户ID为：");
+                messages.append(String.valueOf(userId));
+                messages.append("\n");
+                messages.append("请联系管理员!");
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
             }
         }
 
@@ -52,44 +59,84 @@ public class OrderController extends SimpleListenerHost {
             } else {
                 event.getSubject().sendMessage("不存在指令："+key);
             }
-
         }
 
         //修改指令
         if (Order.ORDER_UPDATE.equals(order)) {
-            String key = split[1];
-            String value = split[2];
-            if (orderMap.get(key) != null) {
-                FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
-                FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
-                event.getSubject().sendMessage("修改成功："+key);
+            if (userId == ConfigData.INSTANCE.getOwner()) {
+                String key = split[1];
+                String value = split[2];
+                if (orderMap.get(key) != null) {
+                    FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
+                    FileUtil.write(ConstantUtil.ORDER_FILENAME,key,value);
+                    event.getSubject().sendMessage("修改成功："+key);
+                } else {
+                    event.getSubject().sendMessage("不存在指令："+key);
+                }
             } else {
-                event.getSubject().sendMessage("不存在指令："+key);
+                MessageChainBuilder messages = new MessageChainBuilder();
+                messages.append("当前用户ID为：");
+                messages.append(String.valueOf(userId));
+                messages.append("\n");
+                messages.append("请联系管理员!");
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
             }
+
         }
 
         //所有指令
         if(Order.ORDER_ALL.equals(order)) {
-            MessageChainBuilder messages = FileUtil.findALL(ConstantUtil.ORDER_FILENAME);
-            MessageChain chain =  messages.build();
-            event.getSubject().sendMessage(chain);
+            if (userId == ConfigData.INSTANCE.getOwner()) {
+                MessageChainBuilder messages = FileUtil.findALL(ConstantUtil.ORDER_FILENAME);
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
+            } else {
+                MessageChainBuilder messages = new MessageChainBuilder();
+                messages.append("当前用户ID为：");
+                messages.append(String.valueOf(userId));
+                messages.append("\n");
+                messages.append("请联系管理员!");
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
+            }
         }
 
         //删除指令
         if (Order.ORDER_REMOVE.equals(order)) {
-            String key = split[1];
-            if (orderMap.get(key) != null) {
-                FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
-                event.getSubject().sendMessage("已删除："+key);
+            if (userId == ConfigData.INSTANCE.getOwner()) {
+                String key = split[1];
+                if (orderMap.get(key) != null) {
+                    FileUtil.remove(ConstantUtil.ORDER_FILENAME,key);
+                    event.getSubject().sendMessage("已删除："+key);
+                } else {
+                    event.getSubject().sendMessage("不存在指令："+key);
+                }
             } else {
-                event.getSubject().sendMessage("不存在指令："+key);
+                MessageChainBuilder messages = new MessageChainBuilder();
+                messages.append("当前用户ID为：");
+                messages.append(String.valueOf(userId));
+                messages.append("\n");
+                messages.append("请联系管理员!");
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
             }
         }
 
         //删除数据
         if (Order.ORDER_DELETE.equals(order)) {
-            FileUtil.deleteOrder(ConstantUtil.ORDER_FILENAME);
-            event.getSubject().sendMessage("数据清除成功！");
+            if (userId == ConfigData.INSTANCE.getOwner()) {
+                FileUtil.deleteOrder(ConstantUtil.ORDER_FILENAME);
+                event.getSubject().sendMessage("数据清除成功！");
+            } else {
+                MessageChainBuilder messages = new MessageChainBuilder();
+                messages.append("当前用户ID为：");
+                messages.append(String.valueOf(userId));
+                messages.append("\n");
+                messages.append("请联系管理员!");
+                MessageChain chain =  messages.build();
+                event.getSubject().sendMessage(chain);
+            }
         }
     }
 }
